Updates the existing test cases.
