+++ tests/UnitTests/Base64DecoderInterfaceTest.php
@@ -3,11 +3,11 @@ namespace CodeKandis\Base64Codec\Tests\UnitTests;
 
 use ArrayIterator;
 use CodeKandis\Base64Codec\Base64DecoderInterface;
-use CodeKandis\Base64Codec\Tests\DataProviders\Base64DecoderInterfaceTest\Base64DecodersWithInvalidBase64EncodedValuesExceptionsErrorCodesAndErrorMessagesDataProvider;
-use CodeKandis\Base64Codec\Tests\DataProviders\Base64DecoderInterfaceTest\Base64DecodersWithInvalidBase64UriSafeEncodedValuesExceptionsErrorCodesAndErrorMessagesDataProvider;
-use CodeKandis\Base64Codec\Tests\DataProviders\Base64DecoderInterfaceTest\Base64DecodersWithValidBase64EncodedValuesAndDecodedValuesDataProvider;
-use CodeKandis\Base64Codec\Tests\DataProviders\Base64DecoderInterfaceTest\Base64DecodersWithValidBase64UriSafeEncodedValuesAndDecodedValuesDataProvider;
-use PHPUnit\Framework\TestCase;
+use CodeKandis\Base64Codec\Tests\DataProviders\UnitTests\Base64DecoderInterfaceTest\Base64DecodersWithInvalidBase64EncodedValuesExceptionsErrorCodesAndErrorMessagesDataProvider;
+use CodeKandis\Base64Codec\Tests\DataProviders\UnitTests\Base64DecoderInterfaceTest\Base64DecodersWithInvalidBase64UriSafeEncodedValuesExceptionsErrorCodesAndErrorMessagesDataProvider;
+use CodeKandis\Base64Codec\Tests\DataProviders\UnitTests\Base64DecoderInterfaceTest\Base64DecodersWithValidBase64EncodedValuesAndDecodedValuesDataProvider;
+use CodeKandis\Base64Codec\Tests\DataProviders\UnitTests\Base64DecoderInterfaceTest\Base64DecodersWithValidBase64UriSafeEncodedValuesAndDecodedValuesDataProvider;
+use CodeKandis\PhpUnit\TestCase;
 
 /**
  * Represents the test case to test objects against the `Base64DecoderInterface`.
